split test_parallel_executor_seresnext test=develop
